Fix build failure
