+++ src/Microsoft.Azure.SignalR.AspNet/EndpointRouters/DefaultEndpointRouter.cs
@@ -8,7 +8,7 @@ using Microsoft.Owin;
 
 namespace Microsoft.Azure.SignalR.AspNet
 {
-    internal class DefaultEndpointRouter : IEndpointRouter
+    internal class DefaultEndpointRouter : DefaultMessageRouter, IEndpointRouter
     {
         /// <summary>
         /// Randomly select from the available endpoints
@@ -21,52 +21,6 @@ namespace Microsoft.Azure.SignalR.AspNet
             // get primary endpoints snapshot
             var availbaleEndpoints = GetNegotiateEndpoints(endpoints);
             return availbaleEndpoints[StaticRandom.Next(availbaleEndpoints.Length)];
-        }
-
-        /// <summary>
-        /// Broadcast to all available endpoints
-        /// </summary>
-        /// <param name="endpoints"></param>
-        /// <returns></returns>
-        public IEnumerable<ServiceEndpoint> GetEndpointsForBroadcast(IEnumerable<ServiceEndpoint> endpoints)
-        {
-            // broadcast to all the endpoints
-            return endpoints;
-        }
-
-        /// <summary>
-        /// Broadcast to all available endpoints
-        /// </summary>
-        /// <param name="userId"></param>
-        /// <param name="endpoints"></param>
-        /// <returns></returns>
-        public IEnumerable<ServiceEndpoint> GetEndpointsForUser(string userId, IEnumerable<ServiceEndpoint> endpoints)
-        {
-            return endpoints;
-        }
-
-        /// <summary>
-        /// Broadcast to all available endpoints
-        /// </summary>
-        /// <param name="groupName"></param>
-        /// <param name="endpoints"></param>
-        /// <returns></returns>
-        public IEnumerable<ServiceEndpoint> GetEndpointsForGroup(string groupName, IEnumerable<ServiceEndpoint> endpoints)
-        {
-            return endpoints;
-        }
-
-        /// <summary>
-        /// Broadcast to all available endpoints, note that this one is only called when the SDK is not able to identify where the connectionId is.
-        /// When the outcoming connectionId happens to be also connected to this app server, SDK can directly send the messages back to that connectionId
-        /// </summary>
-        /// <param name="connectionId"></param>
-        /// <param name="endpoints"></param>
-        /// <returns></returns>
-        public IEnumerable<ServiceEndpoint> GetEndpointsForConnection(string connectionId, IEnumerable<ServiceEndpoint> endpoints)
-        {
-            // broadcast to all the endpoints and service side to do the filter
-            return endpoints;
         }
 
         /// <summary>
