+++ src/Microsoft.Azure.SignalR.Common/EndpointRouters/IMessageRouter.cs
